Use float instead of uint32_t for performance metrics

+++ include/common/Performance.hpp
@@ -4,8 +4,8 @@
 namespace Common {
     namespace Performance {
         struct Frame {
-            uint32_t averageFrameTime;
-            uint32_t elapsedTime;
+            float averageFrameTime;
+            float elapsedTime;
         };
     };
 };
